\name{plot.KLopt.lnorm}
\alias{plot.KLopt.lnorm}
\title{Plot of \eqn{\Psi}{Psi} function for resulting design}
\description{Plots the function from the formulation of the equivalence theorem for resulting approximation \eqn{\xi^{**}}{xi^{**}} of the \eqn{KL}{KL}-optimal design achieved with the help of \code{\link{KLopt.lnorm}}. 
}
\usage{
\method{plot}{KLopt.lnorm}(x, ...)
}
\arguments{
  \item{x}{an object of type "tpopt".}
  \item{\dots}{additional graphical parameters.}
}
\seealso{\code{\link{tpopt}}, \code{\link{summary.tpopt}}, \code{\link{print.tpopt}}}
\keyword{auxiliary}
