\name{print.KLopt.lnorm}
\alias{print.KLopt.lnorm}
\title{Short information about the input}
\description{Prints short information about the input object of class ``KLopt.lnorm''.}
\usage{
\method{print}{KLopt.lnorm}(x, ...)
}
\arguments{
  \item{x}{an object of type "KLopt.lnorm".}
  \item{\dots}{further arguments.}
}
\details{List of models, list of fixed parameters and resulting design are displayed.}
\seealso{\code{\link{KLopt.lnorm}}, \code{\link{summary.KLopt.lnorm}}, \code{\link{plot.KLopt.lnorm}}}
\keyword{auxiliary}