\name{print}
\alias{print.tpopt}
\title{Short information about the input}
\description{Prints short information about the input object of class ``tpopt''.}
\usage{
\method{print}{tpopt}(x, ...)
}
\arguments{
  \item{x}{an object of type "tpopt".}
  \item{\dots}{further arguments.}
}
\details{List of models, list of fixed parameters and resulting design are displayed.}
\seealso{\code{\link{tpopt}}, \code{\link{summary.tpopt}}, \code{\link{plot.tpopt}}}
\keyword{auxiliary}