\name{summary.KLopt.lnorm}
\alias{summary.KLopt.lnorm}
\title{Detailed information about the input}
\description{Prints detailed information about the input object of class ``KLopt.lnorm''.}
\usage{
\method{summary}{KLopt.lnorm}(object, ...)
}
\arguments{
  \item{object}{an object of type "KLopt.lnorm".}
  \item{\dots}{further arguments.}
}
\details{Call, list of models, list of fixed parameters, resulting design, efficiency by iteration and overall execution time are displayed.}
\seealso{\code{\link{KLopt.lnorm}}, \code{\link{plot.KLopt.lnorm}}, \code{\link{print.KLopt.lnorm}}}
\keyword{auxiliary}
