\name{summary.tpopt}
\alias{summary.tpopt}
\title{Detailed information about the input}
\description{Prints detailed information about the input object of class ``tpopt''.}
\usage{
\method{summary}{tpopt}(object, ...)
}
\arguments{
  \item{object}{an object of type "tpopt".}
  \item{\dots}{further arguments.}
}
\details{Call, list of models, list of fixed parameters, resulting design, efficiency by iteration and overall execution time are displayed.}
\seealso{\code{\link{tpopt}}, \code{\link{plot.tpopt}}, \code{\link{print.tpopt}}}
\keyword{auxiliary}
